+++ tex_text/example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+// ignore: depend_on_referenced_packages
 import 'package:tex_text/tex_text.dart';
 
 void main(List<String> args) {
@@ -19,34 +20,62 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(title: const Text("Tex Text.")),
-        body: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              AnimatedBuilder(
-                  animation: _text,
-                  builder: (context, child) {
-                    return TexText(
-                      _text.text,
-                      alignment: TexAlignment.start,
-                      style: Theme.of(context).textTheme.titleLarge?.copyWith(),
-                      mathStyle: MathStyle.text,
-                    );
-                  }),
-              const SizedBox(
-                height: 100,
+        body: Column(
+          mainAxisAlignment: MainAxisAlignment.start,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Expanded(
+              child: ListView(
+                padding: const EdgeInsets.all(10),
+                children: [
+                  AnimatedBuilder(
+                      animation: _text,
+                      builder: (context, child) {
+                        return Material(
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(4),
+                            side: const BorderSide(width: 1),
+                          ),
+                          child: TexText(
+                            // TexText.newEasySyntax(_text.text),
+                            _text.text,
+                            alignment: TexAlignment.start,
+                            style: Theme.of(context)
+                                .textTheme
+                                .headlineLarge
+                                ?.copyWith(
+                                    // color: Colors.red,
+                                    ),
+                            mathStyle: MathStyle.displayCramped,
+                          ),
+                        );
+                      }),
+                ],
               ),
-              TextField(
-                controller: _text,
-                maxLines: null,
-                decoration: const InputDecoration(
-                  border: OutlineInputBorder(),
-                ),
+            ),
+            TextField(
+              controller: _text,
+              maxLines: null,
+              decoration: const InputDecoration(
+                border: OutlineInputBorder(),
               ),
-            ],
-          ),
+            ),
+            Builder(builder: (context) {
+              return FilledButton(
+                onPressed: () {
+                  showDialog(
+                    context: context,
+                    builder: (context) => SimpleDialog(
+                      children: [
+                        SelectableText(TexText.toHtmlData((_text.text)))
+                      ],
+                    ),
+                  );
+                },
+                child: const Text("To html"),
+              );
+            })
+          ],
         ),
       ),
     );